Upload_Kit/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.11.7
motor==3.7.1
pymongo==4.16.0
firebase-admin==7.2.0
razorpay==2.0.0
imagekitio==5.2.0
python-dotenv==1.1.0
python-multipart
